openProfessorApi/requirements.txt
flask==2.3.2
flask-restful==0.3.10
flask-jwt-extended==4.5.2
Flask-SQLAlchemy==3.0.5
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
psycopg2==2.9.6
gunicorn==21.2.0
numpy==1.26.2
